Moved the first few cdk.graph class tests
