+++ base/standard/src/test/java/org/openscience/cdk/graph/AtomContainerAtomPermutorTest.java
@@ -22,13 +22,12 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.openscience.cdk.Atom;
 import org.openscience.cdk.DefaultChemObjectBuilder;
-import org.openscience.cdk.test.CDKTestCase;
 import org.openscience.cdk.interfaces.IAtomContainer;
 import org.openscience.cdk.interfaces.IBond;
 
 /**
  */
-class AtomContainerAtomPermutorTest extends CDKTestCase {
+class AtomContainerAtomPermutorTest {
 
     AtomContainerAtomPermutorTest() {
         super();
