+++ base/standard/src/test/java/org/openscience/cdk/graph/AtomContainerPermutorTest.java
@@ -22,48 +22,45 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.openscience.cdk.Atom;
 import org.openscience.cdk.DefaultChemObjectBuilder;
-import org.openscience.cdk.test.CDKTestCase;
 import org.openscience.cdk.interfaces.IAtomContainer;
 import org.openscience.cdk.interfaces.IBond;
 
-/**
- */
-class AtomContainerBondPermutorTest extends CDKTestCase {
-
-    AtomContainerBondPermutorTest() {
-        super();
-    }
+class AtomContainerPermutorTest {
 
     @Test
-    void constructorTest() {
-        IAtomContainer atomContainer = DefaultChemObjectBuilder.getInstance().newAtomContainer();
-        atomContainer.addAtom(new Atom("C"));
-        atomContainer.addAtom(new Atom("O"));
-        atomContainer.addAtom(new Atom("S"));
-        atomContainer.addBond(0, 1, IBond.Order.SINGLE);
-        atomContainer.addBond(0, 2, IBond.Order.SINGLE);
-        AtomContainerBondPermutor acbp = new AtomContainerBondPermutor(atomContainer);
-        Assertions.assertNotNull(acbp);
-    }
-
-    @Test
-    void containerFromPermutationTest() {
-        IAtomContainer atomContainer = DefaultChemObjectBuilder.getInstance().newAtomContainer();
-        atomContainer.addAtom(new Atom("C"));
-        atomContainer.addAtom(new Atom("O"));
-        atomContainer.addAtom(new Atom("S"));
-        atomContainer.addBond(0, 1, IBond.Order.SINGLE);
-        atomContainer.addBond(0, 2, IBond.Order.SINGLE);
-        AtomContainerBondPermutor acbp = new AtomContainerBondPermutor(atomContainer);
-        IAtomContainer permuted = acbp.containerFromPermutation(new int[]{1, 0, 2});
-        Assertions.assertNotNull(permuted);
-        Assertions.assertEquals(atomContainer.getAtomCount(), permuted.getAtomCount());
-        Assertions.assertEquals(atomContainer.getBondCount(), permuted.getBondCount());
+    void testAtomPermutation() {
+        IAtomContainer ac = DefaultChemObjectBuilder.getInstance().newAtomContainer();
+        IAtomContainer result;
+        String atoms;
+        ac.addAtom(new Atom("C"));
+        ac.addAtom(new Atom("N"));
+        ac.addAtom(new Atom("P"));
+        ac.addAtom(new Atom("O"));
+        ac.addAtom(new Atom("S"));
+        ac.addAtom(new Atom("Br"));
+        ac.addBond(0, 1, IBond.Order.SINGLE);
+        ac.addBond(1, 2, IBond.Order.SINGLE);
+        ac.addBond(2, 3, IBond.Order.SINGLE);
+        ac.addBond(3, 4, IBond.Order.SINGLE);
+        ac.addBond(4, 5, IBond.Order.SINGLE);
+        AtomContainerAtomPermutor acap = new AtomContainerAtomPermutor(ac);
+        int counter = 0;
+        while (acap.hasNext()) {
+            counter++;
+            atoms = "";
+            result = acap.next();
+            for (int f = 0; f < result.getAtomCount(); f++) {
+                atoms += result.getAtom(f).getSymbol();
+            }
+        }
+        Assertions.assertEquals(719, counter);
     }
 
     @Test
     void testBondPermutation() {
         IAtomContainer ac = DefaultChemObjectBuilder.getInstance().newAtomContainer();
+        IAtomContainer result;
+        String bonds;
         ac.addAtom(new Atom("C"));
         ac.addAtom(new Atom("N"));
         ac.addAtom(new Atom("P"));
@@ -79,7 +76,12 @@ class AtomContainerBondPermutorTest extends CDKTestCase {
         int counter = 0;
         while (acap.hasNext()) {
             counter++;
-            acap.next();
+            bonds = "";
+            result = acap.next();
+            for (int f = 0; f < result.getBondCount(); f++) {
+                bonds += result.getBond(f).getOrder();
+            }
+            //logger.debug(bonds);
         }
         Assertions.assertEquals(119, counter);
     }
